+++ Shelf_Edge_Clock_V1/digits.ino
@@ -1,46 +1,46 @@
-void digitZero(int offset, int colour){
+void digitZero(long offset, int colour){
     stripClock.fill(colour, (0 + offset), 27);
     stripClock.fill(colour, (36 + offset), 27);
 }
 
-void digitOne(int offset, int colour){
+void digitOne(long offset, int colour){
     stripClock.fill(colour, (0 + offset), 9);
     stripClock.fill(colour, (36 + offset), 9);
 }
 
-void digitTwo(int offset, int colour){
+void digitTwo(long offset, int colour){
     stripClock.fill(colour, (0 + offset), 18);
     stripClock.fill(colour, (27 + offset), 9);
     stripClock.fill(colour, (45 + offset), 18);
 }
 
-void digitThree(int offset, int colour){
+void digitThree(long offset, int colour){
     stripClock.fill(colour, (0 + offset), 18);
     stripClock.fill(colour, (27 + offset), 27);
 }
 
-void digitFour(int offset, int colour){
+void digitFour(long offset, int colour){
     stripClock.fill(colour, (0 + offset), 9);
     stripClock.fill(colour, (18 + offset), 27);
 }
 
-void digitFive(int offset, int colour){
+void digitFive(long offset, int colour){
     stripClock.fill(colour, (9 + offset), 45);
 }
 
-void digitSix(int offset, int colour){
+void digitSix(long offset, int colour){
     stripClock.fill(colour, (9 + offset), 54);
 }
 
-void digitSeven(int offset, int colour){
+void digitSeven(long offset, int colour){
     stripClock.fill(colour, (0 + offset), 18);
     stripClock.fill(colour, (36 + offset), 9);
 }
 
-void digitEight(int offset, int colour){
+void digitEight(long offset, int colour){
     stripClock.fill(colour, (0 + offset), 63);
 }
 
-void digitNine(int offset, int colour){
+void digitNine(long offset, int colour){
     stripClock.fill(colour, (0 + offset), 45);
 }
